加入录像申请退出等功能

+++ src/main/java/com/randioo/doudizhu_server/module/money/service/MoneyExchangeService.java
@@ -8,7 +8,7 @@ public interface MoneyExchangeService extends ObserveBaseServiceInterface {
 
 	GeneratedMessage moneyExchange(Role role, boolean add, int num);
 	
-	public  boolean exchangeMoney(Role role, int money, boolean add);
+	public  boolean exchangeMoney(Role role, double money, boolean add);
 
 	void newRoleInit(Role role);
